updated the class food to obtain the getter. updated the class EatenFood to obtain the getter. updated the LogicManager to insert String username as parameter of lookUpMostFollowedDietByNutritionist
updated the dietManger: in each if of each possibile operation, i called the corrisponding function of the logicManger (appropiately with the check needed) and the function of the CLI for printing.

updated the CLI class to insert the print of various entitis such as Diet (List of Diet).
MISSING the method for printing Food /User(and the corrisponding List)