inserted the methods for checking and printing the Diet progress and a third method for asking the quantity of the eatenFood in CLI class.
Updated the method checkDietProgress.
Now we consider the total g of each E.Food for each nutrients.
Then we dived the total for each nutrient with the total g of the entire E.Food list to get the average and then we compare these average values with the Diet values.

Then with checkDietProgress() we return if the check is successfully ended ot not. And with the dietProgress(), we return the Average values for each nutrient and the corrisponding value of the Diet.

Tried to print in the correct oreder the System.err with no success.
Then i updated the DietManger according to the updates above.